Initial functioning KeyPromptTexture
The key_prompt_texture_orig.svg file is the icon for KeyPromptTexture before svgo has been used on it, in case it needs to be edited later.
Other assets are from Xelu's Controller Prompt Pack.
Code has not been commented or cleaned up yet.
No functional demo or GH readme.
Not submitted to the asset library yet.
What is working:
- Picking buttons for key prompt textures
- Texture changing when controller changes (only tested w/ Xbox controller)
Not tested: Localization