Optimize texture loading in KeyPromptTexture
Now only the textures for the current controller will be in memory at any given time, as opposed to the previous system of loading them all upon initialization.
This should reduce memory usage and startup times significantly, at the cost of a slight increase in loading demand when the controller changes.
Also fixed a super minor typographigcal error in the code comments.